+++ CHANGELOG.md
@@ -15,6 +15,7 @@ https://gist.github.com/ThornDuke/64da76cd4a56b16492d5101691f6108f
 - package.json: improved `description` for SEO, optimized `keywords` with relevant terms, included
   `sideEffects` for tree-shaking, added `exports` for module compatibility, updated `engines` with
   minimum npm version, and added new `scripts` for a message to the user.
+- index.js: small aesthetic changes
 
 ## [3.0.7] - 2025-04-13